# api/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Criação de uma API com nodeJS, para disponibilizar dados do livro de Monstros do D&D 5e, esse projeto é um trabalho em grupo da materia Arquitetura de computadores, que visa colaborar com a comunidade de RPG de mesa. https://github.com/laurencioScript/api_monster",
    "version": "1.0.0",
    "title": "API Monster D&D"
  },
  "host": "0.0.0.0:8080",
...
  "paths": {
    "/monster/register": {
      "post": {
        "tags": [
          "Monstro"
        ],
        "summary": "",
        "description": "",
        "operationId": "placeOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Cadastra os monstros no banco de dados.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exibe todos os monstros cadastrados"
          },
          "400": {
            "description": " ta bugado me chama laurencio.arkauss@gmail.com "
          }
        }
      }
    },
    "/monster/getMonsterId/": {
      "get": {
        "tags": [
          "Monstro"
        ],
        "summary": "",
        "description": "Retorna todos os dados de um monstro especifico.",
        "operationId": "getMonsterId",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Identificação do monstro",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso na operação",
            "schema": {
              "$ref": "#/definitions/Monstro"
            }
          },
          "404": {
            "description": "ausência de parâmetros"
          }
        }
      }
    },
    "/monster/getMonsterName/": {
      "get": {
        "tags": [
          "Monstro"
        ],
        "summary": "",
        "description": "Retorna todos os dados de um monstro especifico.",
        "operationId": "getMonsterName",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "nome do monstro",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso na operação",
            "schema": {
              "$ref": "#/definitions/Monstro"
            }
          },
          "400": {
            "description": " ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "ausência de parâmetros"
          }
        }
      }
    },
    "/monster/getMonster/": {
      "get": {
        "tags": [
          "Monstro"
        ],
        "summary": "",
        "description": "Busca todos os monstros de acordo com o parametros passados, não deixe espaço vazio substitua por null",
        "operationId": "getMonster",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "description": "tipo do monstro",
            "required": false,
            "type": "string"
          },
          {
            "name": "tamanho",
            "in": "query",
            "description": "tamanho do monstro",
            "required": false,
            "type": "string"
          },
          {
            "name": "desafio",
            "in": "query",
            "description": "desafio do monstro",
            "required": false,
            "type": "string"
          },
          {
            "name": "tendencia",
            "in": "query",
            "description": "tendencia do monstro(sigla em maiusculo)",
            "required": false,
            "type": "string"
          },
          {
            "name": "ca",
            "in": "query",
            "description": "classe de armadura do monstro",
            "required": false,
            "type": "integer"
          },
          {
            "name": "pv",
            "in": "query",
            "description": "pontos de vida do monstro",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Monstros": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "desafio": {
                        "type": "string"
                      },
                      "pv": {
                        "type": "string"
                      },
                      "ca": {
                        "type": "integer"
                      },
                      "tipo": {
                        "type": "string"
                      },
                      "tamanho": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": " ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "ausência de parâmetros"
          }
        }
      }
    },
    "/monster/getAtrib/": {
      "get": {
        "tags": [
          "Monstro"
        ],
        "summary": "",
        "description": "Busca todos os monstros de acordo com o parametros passados, não deixe espaço vazio substitua por null.",
        "operationId": "getAtrib",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "forca",
            "in": "query",
            "description": "valor de força do monstro",
            "required": false,
            "type": "integer"
          },
          {
            "name": "destreza",
            "in": "query",
            "description": "valor de destreza do monstro",
            "required": false,
            "type": "integer"
          },
          {
            "name": "constituicao",
            "in": "query",
            "description": "valor de constituição do monstro",
            "required": false,
            "type": "integer"
          },
          {
            "name": "inteligencia",
            "in": "query",
            "description": "valor de inteligencia do monstro",
            "required": false,
            "type": "integer"
          },
          {
            "name": "sabedoria",
            "in": "query",
            "description": "valor de sabedoria do monstro",
            "required": false,
            "type": "integer"
          },
          {
            "name": "carisma",
            "in": "query",
            "description": "valor de carisma do monstro",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Monstros": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "Força": {
                        "type": "integer"
                      },
                      "Destreza": {
                        "type": "integer"
                      },
                      "Constituição": {
                        "type": "integer"
                      },
                      "Inteligencia": {
                        "type": "integer"
                      },
                      "Sabedoria": {
                        "type": "integer"
                      },
                      "Carisma": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
    "/monster/getDesloc/": {
      "get": {
        "tags": [
          "Monstro"
        ],
        "summary": "",
        "description": "Busca todos os monstros de acordo com o parametros passados, não deixe espaço vazio substitua por null.",
        "operationId": "getDesloc",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "escalada",
            "in": "query",
            "description": "valor de escalada do monstro",
            "required": false,
            "type": "number"
          },
          {
            "name": "voo",
            "in": "query",
            "description": "valor de voo do monstro",
            "required": false,
            "type": "number"
          },
          {
            "name": "escavacao",
            "in": "query",
            "description": "valor de escavação do monstro",
            "required": false,
            "type": "number"
          },
          {
            "name": "natacao",
            "in": "query",
            "description": "valor de natação do monstro",
            "required": false,
            "type": "number"
          },
          {
            "name": "deslocamento",
            "in": "query",
            "description": "valor de deslocamento do monstro",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Monstros": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "Escalada": {
                        "type": "number"
                      },
                      "Voo": {
                        "type": "number"
                      },
                      "Escavação": {
                        "type": "number"
                      },
                      "Natação": {
                        "type": "number"
                      },
                      "Deslocamento": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
    "/monster/getPericia/": {
      "get": {
        "tags": [
          "Monstro"
        ],
        "summary": "",
        "description": "Busca todos os monstros de acordo com o parametros passados, não deixe espaço vazio substitua por null.",
        "operationId": "getPericia",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "parm1",
            "in": "query",
            "description": "pericia do monstro",
            "required": false,
            "type": "string"
          },
          {
            "name": "parm2",
            "in": "query",
            "description": "pericia do monstro",
            "required": false,
            "type": "string"
          },
          {
            "name": "parm3",
            "in": "query",
            "description": "pericia do monstro",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Monstros": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "pericia": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "tipo": {
                              "type": "string"
                            },
                            "bonus": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
    "/monster/getIdioma/": {
      "get": {
        "tags": [
          "Monstro"
        ],
        "summary": "",
        "description": "Busca todos os monstros de acordo com o parametros passados, não deixe espaço vazio substitua por null.",
        "operationId": "getIdioma",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "parm1",
            "in": "query",
            "description": "idioma do monstro",
            "required": false,
            "type": "string"
          },
          {
            "name": "parm2",
            "in": "query",
            "description": "idioma do monstro",
            "required": false,
            "type": "string"
          },
          {
            "name": "parm3",
            "in": "query",
            "description": "idioma do monstro",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Monstros": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "idioma": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Register": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "desafio": {
            "type": "integer"
          },
          "xp": {
            "type": "integer"
          },
          "nome": {
            "type": "string"
          },
          "tipo": {
            "type": "string"
          },
          "tamanho": {
            "type": "string"
          },
          "tendencia": {
            "type": "string"
          },
          "pv": {
            "type": "string"
          },
          "ca": {
            "type": "integer"
          }
        }
      }
    },
...
  },
...
}